Move protocol tests into a separate binary
This ensures all test output is properly captured, avoids potential
bugs in Vulkan libraries, drivers, and layers if multiple Vulkan instances
are created and destroyed by different threads, allows for tests to be
skipped at runtime if prerequisites do not hold, and will make it possible
in the future to test different builds of Waypipe against each other.

The cost is a lot of extra build and test framework complexity.