Define a more correct stat for FreeBSD
The libc crate's struct stat definition is incorrect on FreeBSD >= 12.0,
so redefine the parts of it that Waypipe uses.

Waypipe currently needs FreeBSD >= 13.0 for eventfd and memfd_create
support and so does not yet need to support the pre-12.0 stat layout;
although this could be added in the future.

The code directly calling the C standard library's stat() has not yet
been tested on FreeBSD, but works on Linux with modified values for
sizeof/offsetof.