screencopy: Track copy_with_damage
If waypipe does not recognize the buffer being attached with a copy
request, it will error when the ready event is received as it does not
know which buffer the operation is performed on.

Track the copy_with_damage event to know which buffer is used when the
client wants to wait for damage to accumulate before capturing a frame.