Warn if submitting bad modifier for ext-image-copy-capture
Waypipe currently does not have a way to ensure a dmabuf created
on the display side for a wl_buffer will have a modifier from
a sublist of the set of all valid modifiers, and may thus submit
dmabufs not supported by ext-image-copy-capture-v1. Printing a warning
when this occurs is better than nothing.