Restrict dmabufs to use screencopy-safe modifiers
When ext_image_copy_capture_session_v1 communicates a list of
acceptable format modifier pairs, Waypipe has no way to determine,
when it creates a DMABUF for a wl_buffer, whether the DMABUF
will be used for an ext_image_copy_capture_frame_v1 or not. To make
it more likely that the modifiers will work for any role, all
future DMABUFs created using a format ever advertised as a
screencopy target will be restricted to using modifiers that
were once stated to be acceptable.

This is a hack, but may work OK in practice. Protocols like
zwlr-export-dmabuf where the compositor creates the images
are unaffected.