Support 32 bit systems with 64 bit time
On some 32 bit systems with 64 bit time, including Debian's armhf after the t64
transition, libc::timespec is not directly constructible as a struct literal due
to the possible presence of (private) padding. Work around that by initializing
the struct to zero, and set the desired fields.

See also the libc crate's issue 3223 [1].

[1] https://github.com/rust-lang/libc/issues/3223